relocate files
