--- api/common/__init__.py
@@ -1,9 +1,0 @@
-"""
-© Ocado Group
-Created on 13/09/2024 at 12:00:14(+03:00).
-"""
-
-from .model_serializer import ModelSerializer
-from .model_serializer_test_case import ModelSerializerTestCase
-from .model_view_set import ModelViewSet
-from .model_view_set_test_case import ModelViewSetTestCase
