+++ api/serializers/_model_list_serializer.py
@@ -6,20 +6,13 @@ Created on 13/09/2024 at 12:00:41(+03:00).
 import typing as t
 
 from codeforlife.request import BaseRequest
-from codeforlife.serializers import BaseModelListSerializer, BaseModelSerializer
+from codeforlife.serializers import BaseModelListSerializer
 from django.db.models import Model
 
 from ..models import Contributor
 from ..models.session import SessionStore
 
 AnyModel = t.TypeVar("AnyModel", bound=Model)
-
-
-class ModelSerializer(
-    BaseModelSerializer[BaseRequest[SessionStore, Contributor], AnyModel],
-    t.Generic[AnyModel],
-):
-    """Base model serializer."""
 
 
 class ModelListSerializer(
