+++ api/serializers/_model_serializer_test_case.py
@@ -6,39 +6,18 @@ Created on 13/09/2024 at 12:00:34(+03:00).
 import typing as t
 
 from codeforlife.request import BaseRequest
-from codeforlife.tests import (
-    BaseAPIRequestFactory,
-    BaseModelListSerializerTestCase,
-    BaseModelSerializerTestCase,
-)
+from codeforlife.tests import BaseAPIRequestFactory, BaseModelSerializerTestCase
 from django.db.models import Model
 
 from ..models import Contributor
 from ..models.session import SessionStore
-from .model_serializer import ModelListSerializer, ModelSerializer
+from ._model_serializer import ModelSerializer
 
 AnyModel = t.TypeVar("AnyModel", bound=Model)
 
 
 class ModelSerializerTestCase(
     BaseModelSerializerTestCase[
-        ModelSerializer[AnyModel],
-        BaseAPIRequestFactory[
-            BaseRequest[SessionStore, Contributor], Contributor
-        ],
-        AnyModel,
-    ],
-    t.Generic[AnyModel],
-):
-    """Base model serializer test case."""
-
-    request_factory_class = BaseAPIRequestFactory
-
-
-# pylint: disable-next=too-many-ancestors
-class ModelListSerializerTestCase(
-    BaseModelListSerializerTestCase[
-        ModelListSerializer[AnyModel],
         ModelSerializer[AnyModel],
         BaseAPIRequestFactory[
             BaseRequest[SessionStore, Contributor], Contributor
