+++ api/serializers/agreement_signature.py
@@ -8,8 +8,8 @@ from datetime import datetime
 from django.utils import timezone
 from rest_framework import serializers
 
-from ..common import ModelSerializer
 from ..models import AgreementSignature
+from ._model_serializer import ModelSerializer
 
 # pylint: disable=missing-class-docstring
 # pylint: disable=missing-function-docstring
